+++ node/directives/checkAdminAccess.ts
@@ -4,6 +4,7 @@ import type { GraphQLField } from 'graphql'
 import { defaultFieldResolver } from 'graphql'
 
 import sendAuthMetric, { AuthMetric } from '../metrics/auth'
+import { validateAdminToken, validateApiToken } from './helper'
 
 export class CheckAdminAccess extends SchemaDirectiveVisitor {
   public visitFieldDefinition(field: GraphQLField<any, any>) {
@@ -16,68 +17,70 @@ export class CheckAdminAccess extends SchemaDirectiveVisitor {
       info: any
     ) => {
       const {
-        vtex: { adminUserAuthToken, logger },
-        clients: { identity },
+        vtex: { adminUserAuthToken, storeUserAuthToken, logger },
       } = context
 
+      const { hasAdminToken, hasValidAdminToken, hasCurrentValidAdminToken } =
+        await validateAdminToken(context, adminUserAuthToken as string)
+
+      const { hasApiToken, hasValidApiToken } = await validateApiToken(context)
+
+      const hasStoreToken = !!storeUserAuthToken // we don't need to validate store token
+
+      // now we emit a metric with all the collected data before we proceed
       const operation = field.astNode?.name?.value ?? context.request.url
-      const metric = new AuthMetric(
-        context.vtex.account,
+      const userAgent = context?.request?.headers['user-agent'] as string
+      const caller = context?.request?.headers['x-vtex-caller'] as string
+      const forwardedHost = context?.request?.headers[
+        'x-forwarded-host'
+      ] as string
+
+      const auditMetric = new AuthMetric(
+        context?.vtex?.account,
         {
           operation,
-          forwardedHost: context.request.header['x-forwarded-host'] as string,
-          caller: context.request.header['x-vtex-caller'] as string,
-          userAgent: context.request.header['user-agent'] as string,
-          hasAdminToken: !!adminUserAuthToken,
-          hasStoreToken: false,
-          hasApiToken: false,
+          forwardedHost,
+          caller,
+          userAgent,
+          hasAdminToken,
+          hasValidAdminToken,
+          hasApiToken,
+          hasValidApiToken,
+          hasStoreToken,
         },
-        'CheckAdminAccess'
+        'CheckAdminAccessAudit'
       )
 
-      if (!adminUserAuthToken) {
-        metric.error = 'No admin token provided'
-        sendAuthMetric(logger, metric)
+      sendAuthMetric(logger, auditMetric)
+
+      if (!hasAdminToken) {
         logger.warn({
-          message: 'CheckAdminAccess: No admin token provided',
-          userAgent: context.request.header['user-agent'],
-          vtexCaller: context.request.header['x-vtex-caller'],
-          forwardedHost: context.request.header['x-forwarded-host'],
+          message: 'CheckAdminAccess: No token provided',
+          userAgent,
+          caller,
+          forwardedHost,
           operation,
+          hasAdminToken,
+          hasValidAdminToken,
+          hasApiToken,
+          hasValidApiToken,
+          hasStoreToken,
         })
         throw new AuthenticationError('No token was provided')
       }
 
-      try {
-        const authUser = await identity.validateToken({
-          token: adminUserAuthToken,
-        })
-
-        // This is the first step before actually enabling this code.
-        // For now we only log in case of errors, but in follow up commits
-        // we should also throw an exception inside this if in case of errors
-        if (!authUser?.audience || authUser?.audience !== 'admin') {
-          metric.error = 'Token is not an admin token'
-          sendAuthMetric(logger, metric)
-          logger.warn({
-            message: `CheckUserAccess: Token is not an admin token`,
-            userAgent: context.request.header['user-agent'],
-            vtexCaller: context.request.header['x-vtex-caller'],
-            forwardedHost: context.request.header['x-forwarded-host'],
-            operation,
-          })
-        }
-      } catch (err) {
-        metric.error = 'Invalid token'
-        sendAuthMetric(logger, metric)
+      if (!hasCurrentValidAdminToken) {
         logger.warn({
-          error: err,
           message: 'CheckAdminAccess: Invalid token',
-          userAgent: context.request.header['user-agent'],
-          vtexCaller: context.request.header['x-vtex-caller'],
-          forwardedHost: context.request.header['x-forwarded-host'],
+          userAgent,
+          caller,
+          forwardedHost,
           operation,
-          token: adminUserAuthToken,
+          hasAdminToken,
+          hasValidAdminToken,
+          hasApiToken,
+          hasValidApiToken,
+          hasStoreToken,
         })
         throw new ForbiddenError('Unauthorized Access')
       }